+++ app/(content)/dashboard/page.tsx
@@ -3,11 +3,18 @@
 import { useEffect, useState } from "react";
 import InnerHeader from "@/components/pagetemplates/innerheader/InnerHeader";
 import { useSession } from "next-auth/react";
-import DashChild from "@/components/pagecomponents/dashboard/dashChild";
 import MainChild from "@/components/pagetemplates/mainchild/mainchild";
 import { IHobby } from "@/models/types/hobby";
 import { useHobbyContext } from "@/app/context/hobby/hobbyModalContext";
 import { useStateContext } from "@/app/context/state/StateContext";
+import DashActionsButton from "@/components/pagecomponents/dashboard/buttons/dashactions";
+import ModalHobby from "@/components/modals/hobbyModal/hobbymodal";
+import DashFilterButton from "@/components/pagecomponents/dashboard/buttons/dashFilter";
+import CalView from "@/components/pagecomponents/dashboard/views/calView";
+import StatsView from "@/components/pagecomponents/dashboard/views/statsView";
+import { useModalContext } from "@/app/context/modal/modalContext";
+import useMediaQuery from "@/components/listeners/WidthSettings";
+import { FiMenu } from "react-icons/fi";
 
 
 export default function Dashboard() {
@@ -16,10 +23,12 @@ export default function Dashboard() {
     const [titles, setTitles] = useState([] as string[]);
     const { loading, setLoading } = useStateContext();
     const { data: session } = useSession();
-    const { hobbies, setHobbies, setRefreshKey, refreshKey } = useHobbyContext();
+    const { hobbies, setHobbies, openAddModal, filterItem, refreshKey } = useHobbyContext();
     const { urlToUse } = useStateContext();
     const adminID = session?.user?.email === process.env.NEXT_PUBLIC_ADMIN_USERNAME ? true : false;
     const userID = process.env.NEXT_PUBLIC_ADMIN_USERNAME;
+    const isBreakpoint = useMediaQuery(768);
+    const { setOpenDashboardMobileDropdown, calDash } = useModalContext();
 
     useEffect(() => {
       const getHobbies =  async () => {
@@ -74,11 +83,9 @@ export default function Dashboard() {
     
 
     return (
-        <div>
+        <div className="h-full w-full">
             <InnerHeader>
-              <div className="flex flex-col justify-end">
-                 <h1 className={`text-xl md:text-4xl font-bold`}>Dashboard</h1>
-              </div>
+                <h1 className={`text-lg md:text-xl font-bold`}>Dashboard</h1>
             </InnerHeader>
             <MainChild>
               {loading ? (
@@ -86,7 +93,29 @@ export default function Dashboard() {
                     <h1>Loading...</h1>
                   </div>
                 ) : (
-                  <DashChild categories={categories} titles={titles} hobbies={hobbies} adminID={adminID}/>
+                  <div className="p-2 h-full">
+                    {!isBreakpoint &&
+                      <div className={`flex flex-col md:flex-row top-0 z-20 items-start md:items-center justify-between`}>
+                        <DashFilterButton titles={titles} categories={categories} />
+
+                        {session?.user !== null && adminID === true &&
+                            <DashActionsButton />
+                        } 
+                      </div>
+                    }
+                    {isBreakpoint &&
+                        <FiMenu className="flex flex-row justify-start m-2" onClick={() => setOpenDashboardMobileDropdown(true)} />
+                    }
+                    <ModalHobby show={openAddModal} categories={categories} hobbies={hobbies} />
+                    {calDash &&
+                      <div className="flex flex-row justify-center items-center" >
+                        <CalView filter={filterItem} hobbies={hobbies} />
+                      </div>
+                    }
+                    {!calDash &&
+                      <StatsView hobbies={hobbies} daysThisMonth={30} />
+                    }
+                  </div>
                 )
               }
             </MainChild>
