+++ components/nav/Navbar.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import SideNavHeader from '@/components/nav/sideNav/header';
+import SideNavHeader from '@/components/nav/header/header';
 import { usePathname } from 'next/navigation';
 
 
@@ -8,7 +8,7 @@ const NavBar = () => {
     const pathName = usePathname();
 
     return (
-        <div className={`flex flex-row h-24 items-center w-full px-10`} style={{height: '8vh'}}>
+        <div className={`flex flex-row h-24 items-center w-full px-5`} style={{border: 'box-border'}}>
           {pathName !== 'demo' &&
           <SideNavHeader/>
           }
