+++ components/nav/header/header.tsx
@@ -4,15 +4,15 @@ import { usePathname } from "next/navigation";
 import { useState, useCallback } from "react";
 import React from "react";
 import useMediaQuery from "@/components/listeners/WidthSettings";
-import Image from "next/image";
-import Sidenav from "./SideNav";
+import Sidenav from "../sideNav/SideNav";
 import { SideMenuAccordian } from "../../dropdowns/SideMenuAccordian";
 import SocialButton from "@/components/buttons/socialButton";
+import ProfilePicture from "./profilepicture";
 
 
 const SideNavHeader = () => {
     const isBreakpoint = useMediaQuery(768);
-    const [open, setOpen] = useState(false); //make sure no dups
+    const [open, setOpen] = useState(false);
     const pathname = usePathname();
     const toggle = () => {
       setOpen((prevState) => !prevState);
@@ -48,47 +48,33 @@ const SideNavHeader = () => {
     return (
       <>
         {pathname !== '/'  && pathname !== '/demo_303' &&
-        <>
+        <div className={`flex flex-row ${isBreakpoint ? 'justify-start' : 'justify-between'} items-center pb-2 w-full`}>
         {isBreakpoint &&
         <>
-          <div className={`flex flex-row justify-start items-center`}>
-          <button 
-            type='button'
-            aria-disabled={open}
-            disabled={open}
-            onClick={toggle}
-            className={`text-black font-medium ${open ? 'text-transparent' : 'text-black'}`}
-          >
-            Menu
-          </button>
-          <>
-          <div className={`mx-5 my-1 ${open ? 'text-transparent' : 'text-black'}`}>|</div>
-          <div>
-            <Link className={`text-black font-medium ${open ? 'text-transparent' : 'text-black'}`} href='/'>
-              Home
-            </Link>
-          </div>
-          </>
-        </div>
-        <div ref={imageRef} className={`mt-3 ml-5 ${clicked ? style.profilepicture.large : style.profilepicture.small}`}>
-        {pathname === '/about/professional' &&
-          <Image
-            onClick={imageClick}
-            priority
-            src="/images/carlseaholmimage.jpg"
-            className={`z-30 rounded-full overflow-x-hidden transition-all ease duration-200 ${isHovered ? 'cursor-pointer' : ''}`}
-            height={clicked ? 200 : 60}
-            width={clicked ? 200 : 60}
-            alt="Carl Seaholm Profile Photo"
-            onMouseEnter={() => setIsHovered(true)}
-            onMouseLeave={() => setIsHovered(false)}
-          />
-        }
-        </div>
+            <button 
+              type='button'
+              aria-disabled={open}
+              disabled={open}
+              onClick={toggle}
+              className={`text-black font-medium ${open ? 'text-transparent' : 'text-black'}`}
+            >
+              Menu
+            </button>
+            <div className={`mx-5 my-1 ${open ? 'text-transparent' : 'text-black'}`}>
+              |
+            </div>
+            <div>
+              <Link className={`text-black font-medium ${open ? 'text-transparent' : 'text-black'}`} href='/'>
+                Home
+              </Link>
+            </div>
+          {pathname === '/about/professional' &&
+            <ProfilePicture style={style.profilepicture} imageClick={imageClick} imageRef={imageRef} isHovered={false} setIsHovered={undefined} clicked={false} />
+          }
         </>
         }
         {!isBreakpoint &&
-          <div className="flex flex-row justify-between items-center w-full">
+          <>
             <div className={`flex flex-row items-center`}>
               <button 
                 type='button'
@@ -113,15 +99,15 @@ const SideNavHeader = () => {
               <p className='mx-3'>|</p>
               <SocialButton networkName='linkedin' parent={true} />
             </div>
-          </div>
-          }
+          </>
+        }
 
           <Sidenav open={open} toggle={toggle}>
             {open ? (
                 <SideMenuAccordian toggle={toggle} />
                 ) : null}
           </Sidenav>
-        </>
+        </div>
         }
       </>
     );
