+++ components/pagetemplates/innerheader/InnerHeader.tsx
@@ -5,7 +5,7 @@ const InnerHeader = ({children}: {children: React.ReactNode;}) => {
     const lengthOfChildren = React.Children.count(children);
 
     return (
-        <div className={`flex flex-row ${lengthOfChildren > 1 ? 'justify-between' : 'justify-center'} items-center w-full p-2 pb-5`} style={{maxHeight: '8vh', minHeight: '8vh'}}>
+        <div className={`flex flex-row ${lengthOfChildren > 1 ? 'justify-between' : 'justify-center'} items-center w-full p-1`}>
             {children}
         </div>
     );
